renamed subset_polygon
