+++ flyingpigeon/processes/__init__.py
@@ -1,5 +1,5 @@
 # from .wps_say_hello import SayHello
-from .wps_subset_polygon import SubsetWFSPolygonProcess
+from .wps_subset_wfs_polygon import SubsetWFSPolygonProcess
 from .wps_subset_bbox import SubsetBboxProcess
 from .wps_subset_continents import SubsetcontinentProcess
 from .wps_subset_countries import SubsetcountryProcess
