+++ tests/test_wps_caps.py
@@ -1,6 +1,4 @@
 from pywps import Service
-from pywps.tests import assert_response_success
-
 from .common import client_for
 from flyingpigeon.processes import processes
